// 3appmovie-app/src/app/core/models/moviehttp.model.ts
export interface MovieModelHttp {

    id: string,
    creationTime: string,
    creatorId: string,
    lastModificationTime: string,
    lastModifierId:string,
    movieAuthor: string,
    title:string,
    genreDescription: string,
    genreid: string,
    authorid: string,
    releaseDate: string,
    authorName: string,
    price: number,
    rating:number,
    runtimeMins: string
    plot: string
    writers: string
    stars: string
    countries: string
    languages: string
    idImdb: string
    awards: string
    linkEmbed: string
    videoDescription: string
    linkPoster1: string
    linkPoster2: string
    image1: string
    image2: string

}